+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - add plugin list modal for updating to hb v2.0
 - change placement of page header tooltips to bottom
+- single modal for showing/hiding widgets
 
 ### Other Changes